+++ apps/server/src/api/subscription/commands/delete-subscription.command.ts
@@ -1,6 +1,6 @@
 import { SubscriptionModel } from "../models/subscription.model"
 
-export async function deleteSubscriptionCommand(workspaceId: string) {
-  const subscription = await SubscriptionModel.findOneAndDelete({ workspaceId })
+export async function deleteSubscriptionCommand(userId: string) {
+  const subscription = await SubscriptionModel.findOneAndDelete({ userId })
   return subscription
 }